+++ common_blocks/pipelines.py
@@ -4,62 +4,15 @@ from steppy.base import Step, IdentityOperation
 from steppy.adapter import Adapter, E
 
 from . import loaders
-from .models import PyTorchUNet
 from .utils import make_apply_transformer
 from .postprocessing import crop_image, resize_image, binarize
-from .pipeline_config import ORIGINAL_SIZE
-
-
-def unet(config, suffix='', train_mode=True):
-    if train_mode:
-        preprocessing = preprocessing_train(config, model_name='unet', suffix=suffix)
-    else:
-        preprocessing = preprocessing_inference(config, suffix=suffix)
-
-    unet = Step(name='unet{}'.format(suffix),
-                transformer=PyTorchUNet(**config.model['unet']),
-                input_data=['callback_input'],
-                input_steps=[preprocessing],
-                adapter=Adapter({'datagen': E(preprocessing.name, 'datagen'),
-                                 'validation_datagen': E(preprocessing.name, 'validation_datagen'),
-                                 'meta_valid': E('callback_input', 'meta_valid'),
-                                 }),
-                is_trainable=True,
-                force_fitting=True,
-                experiment_directory=config.env.experiment_dir)
-    return unet
-
-
-def unet_tta(config, suffix=''):
-    preprocessing, tta_generator = preprocessing_inference_tta(config, model_name='unet')
-
-    unet = Step(name='unet{}'.format(suffix),
-                transformer=PyTorchUNet(**config.model['unet']),
-                input_data=['callback_input'],
-                input_steps=[preprocessing],
-                is_trainable=True,
-                experiment_directory=config.env.experiment_dir)
-
-    tta_aggregator = aggregator('tta_aggregator{}'.format(suffix), unet,
-                                tta_generator=tta_generator,
-                                experiment_directory=config.env.experiment_dir,
-                                config=config.tta_aggregator)
-
-    prediction_renamed = Step(name='prediction_renamed{}'.format(suffix),
-                              transformer=IdentityOperation(),
-                              input_steps=[tta_aggregator],
-                              adapter=Adapter({'mask_prediction': E(tta_aggregator.name, 'aggregated_prediction')
-                                               }),
-                              experiment_directory=config.env.experiment_dir)
-
-    return prediction_renamed
 
 
 def preprocessing_train(config, model_name='unet', suffix=''):
-    if config.execution.loader_mode == 'crop_and_pad':
+    if config.general.loader_mode == 'crop_and_pad':
         Loader = loaders.ImageSegmentationLoaderCropPad
         loader_config = config.loaders.crop_and_pad
-    elif config.execution.loader_mode == 'resize':
+    elif config.general.loader_mode == 'resize':
         Loader = loaders.ImageSegmentationLoaderResize
         loader_config = config.loaders.resize
     else:
@@ -106,10 +59,10 @@ def preprocessing_train(config, model_name='unet', suffix=''):
 
 
 def preprocessing_inference(config, model_name='unet', suffix=''):
-    if config.execution.loader_mode == 'crop_and_pad':
+    if config.general.loader_mode == 'crop_and_pad':
         Loader = loaders.ImageSegmentationLoaderCropPad
         loader_config = config.loaders.crop_and_pad
-    elif config.execution.loader_mode == 'resize':
+    elif config.general.loader_mode == 'resize':
         Loader = loaders.ImageSegmentationLoaderResize
         loader_config = config.loaders.crop_and_pad
     else:
@@ -172,10 +125,10 @@ def preprocessing_inference_tta(config, model_name='unet', suffix=''):
     else:
         raise NotImplementedError
 
-    if config.execution.loader_mode == 'crop_and_pad':
+    if config.general.loader_mode == 'crop_and_pad':
         Loader = loaders.ImageSegmentationLoaderCropPadTTA
         loader_config = config.loader.crop_and_pad_tta
-    elif config.execution.loader_mode == 'resize':
+    elif config.general.loader_mode == 'resize':
         Loader = loaders.ImageSegmentationLoaderResizeTTA
         loader_config = config.loader.resize_tta
     else:
@@ -205,10 +158,10 @@ def aggregator(name, model, tta_generator, experiment_directory, config):
 
 
 def mask_postprocessing(config, suffix=''):
-    if config.execution.loader_mode == 'crop_and_pad':
-        size_adjustment_function = partial(crop_image, target_size=ORIGINAL_SIZE)
-    elif config.execution.loader_mode == 'resize':
-        size_adjustment_function = partial(resize_image, target_size=ORIGINAL_SIZE)
+    if config.general.loader_mode == 'resize_and_pad':
+        size_adjustment_function = partial(crop_image, target_size=config.general.original_size)
+    elif config.general.loader_mode == 'resize':
+        size_adjustment_function = partial(resize_image, target_size=config.general.original_size)
     else:
         raise NotImplementedError
 
@@ -230,10 +183,3 @@ def mask_postprocessing(config, suffix=''):
                                       }),
                      experiment_directory=config.env.experiment_dir)
     return binarizer
-
-
-PIPELINES = {'unet': {'network': unet,
-                      'postprocessing': mask_postprocessing},
-             'unet_tta': {'network': unet_tta,
-                          'postprocessing': mask_postprocessing},
-             }
